Ship

+++ core/src/com/mygdx/game/utils/Vector2D.java
@@ -6,8 +6,8 @@ package com.mygdx.game.utils;
 
 public class Vector2D
 {
-    public double vX;
-    public double vY;
+    public float vX;
+    public float vY;
 
     Vector2D()
     {
@@ -15,7 +15,7 @@ public class Vector2D
         vX=1;
         vY=1;
     }
-    Vector2D(double x,double y)
+    Vector2D(float x,float y)
     {
 
         vX=x;
